OJ-2880: Implement the nonUKAddress form
Implemented the enter-non-UK-address.njk view which layout
the non UK Address entry form using the custom linked input component
along with the address-year-from-field macro

Implements the nonUKAddressValidator which has logic to check the
grouped inputs if they are empty indication validation errors for top area
of the form i.e. buildAddress.

The nonUKAddress controller manages the running of the validation logic i.e.
uses the nonUKAddressValidator if every buildAddress input is found to be empty
the nonUKAddressValidator -> validateBuildingAddress function is added to
validation array. buildAddressErrorMessage is display with input highlighted in red
no message is displayed for inputs themselves see: nonBuildingAddressInputErrors

Whenever the continue button is click the controller runs
all form validation logic for all fields errors are displayed based on
req?.form?.errors

nonUKAddress also manages the data input saving and loading the data as
appropriate and ensuring data is valid before forwarding to confirmation
screen when the continue button is click

Co-authored-by: Mike Beeb <mike.beeby@digital.cabinet-office.gov.uk>
Co-authored-by: Catlin Coolin <caitlin.cooling@digital.cabinet-office.gov.uk>